feat: [FN-316] 그룹 상세 페이지에 가입 및 초대 버튼 기능 추가
- 비멤버에게 가입 버튼 노출 (applicationRequired 여부에 따라 분기)
  - applicationRequired true: 가입신청 다이얼로그 (GroupJoinDialog)
  - applicationRequired false: 즉시 가입 처리
- GroupInviteDialog 컴포넌트 추출 후 멤버 초대 버튼에 연결
- GroupInvitationManagement 인라인 다이얼로그를 GroupInviteDialog로 교체
